Centralize CLI flag handling and create common options across set/get
This isn't as clean as adding a helper method that injects a username
or manta attributes for the get/set sub-commands, but this is the
most clean way in the short term.  If these options are unspecified
globally, they just aren't used.  From a UX perspective, when
introducing a config file, this is probably the cleanest outcome until
the flag/config registration is injected via a helper function.